requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
attr==0.3.2
attrs==24.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
DateTime==5.5
editor==1.6.6
emoji==2.14.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib==1.0.4
inquirer==3.4.0
multidict==6.1.0
munch==4.0.0
outcome==1.3.0.post0
pillow==11.0.0
propcache==0.2.0
pycryptodome==3.21.0
pycurl==7.45.3
PySocks==1.7.1
pystyle==2.9
python-dotenv==1.0.1
pytz==2024.2
readchar==4.2.1
requests==2.32.3
runs==1.2.2
selenium==4.26.1
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wheel==0.45.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.17.1
youtube-search-python==1.6.6
zlapi==1.0.2
zope.interface==7.1.1
